requirements.txt
asgiref==3.6.0
autopep8==2.0.1
bcrypt==4.0.1
Django==3.2
djangorestframework==3.14.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
